backend/requirements-prod.txt
# Production dependencies only — no audio analysis tools
# Essentia, librosa, numba etc. are local-only for track analysis

annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
numpy==2.4.2
packaging==26.0
platformdirs==4.9.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
